requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
discord==2.3.2
discord.py==2.3.2
fonttools==4.43.1
frozendict==2.3.8
frozenlist==1.4.0
html5lib==1.1
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.1
mplcyberpunk==0.7.0
multidict==6.0.4
multitasking==0.0.11
numpy==1.26.1
packaging==23.2
pandas==2.1.2
peewee==3.17.0
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
yarl==1.9.2
yfinance==0.2.31
